+++ docs/release_notes.rst
@@ -9,3 +9,4 @@ Substeps Core - Release Notes
 - Changes to support the eclipse plugin.
 - DefaultExecutionReportBuilder will now delete the report directory if it's already there.
 - substeps filename now reported in debug
+- changes to support glossary building